safari-600.7-branch/LayoutTests/platform/mac/accessibility/aria-directory-expected.txt

Item 1
Item 2
Item 3
This tests that ARIA directory is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dir.role is 'AXRole: AXList'
PASS dir.subrole is 'AXSubrole: AXContentList'
PASS successfullyParsed is true

TEST COMPLETE
